% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAm.R
\name{dnamReport}
\alias{dnamReport}
\title{Generate DNAm PDF Report from DNAm.Rmd - inst/scripts/.R}
\usage{
dnamReport(
  output = "DNAm_Report.pdf",
  outputDir = "reports",
  qcDir = "figures/preprocessingMinfiEwasWater/enMix",
  preprocessingDir = "figures/preprocessingMinfiEwasWater/qc",
  postprocessingDir = "figures/preprocessingMinfiEwasWater/metrics",
  svaDir = "figures/svaEnmix/sva",
  glmDir = "figures/methylationGLM_T1",
  glmmDir = "figures/methylationGLMM_T1T2",
  reportTitle = "DNA methylation",
  author = "School of Biomedical Sciences",
  date = format(Sys.Date(), "\%B \%d, \%Y")
)
}
\description{
Generate DNAm PDF Report from DNAm.Rmd - inst/scripts/.R
}
