% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylationGLM_T1.R
\name{methylationGLM_T1}
\alias{methylationGLM_T1}
\title{Run methylationGLM_T1.R as an external script from inst/script/.R}
\usage{
methylationGLM_T1(
  inputPheno = "rData/preprocessingPheno/mergeData/phenoBetaT1.RData",
  outputLogs = "logs",
  outputRData = "rData/methylationGLM_T1/models",
  outputPlots = "figures/methylationGLM_T1",
  phenotypes =
    "DASS_Depression,DASS_Anxiety,DASS_Stress,PCL5_TotalScore,MHCSF_TotalScore,BRS_TotalScore",
  covariates = "Sex,Age,Ethnicity,TraumaDefinition,Leukocytes,Epithelial.cells",
  factorVars = "Sex,Ethnicity,TraumaDefinition",
  cpgPrefix = "cg",
  cpgLimit = NA,
  nCores = 32,
  plotWidth = 2000,
  plotHeight = 1000,
  plotDPI = 150,
  interactionTerm = NULL,
  libPath = NULL,
  glmLibs = "glm2",
  prsMap = NULL,
  summaryPval = NA,
  summaryResidualSD = TRUE,
  saveSignificantCpGs = FALSE,
  significantCpGDir = "preliminaryResults/cpgs/methylationGLM_T1",
  significantCpGPval = 0.05,
  saveTxtSummaries = TRUE,
  chunkSize = 10000,
  summaryTxtDir = "preliminaryResults/summary/methylationGLM_T1/glm",
  fdrThreshold = 0.05,
  padjmethod = "fdr",
  annotationPackage = "IlluminaHumanMethylationEPICv2anno.20a1.hg38",
  annotationCols =
    "Name,chr,pos,UCSC_RefGene_Group,UCSC_RefGene_Name,Relation_to_Island,GencodeV41_Group",
  annotatedGLMOut = "data/methylationGLM_T1"
)
}
\description{
Run methylationGLM_T1.R as an external script from inst/script/.R
}
