% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylationGLMM_T1T2.R
\name{methylationGLMM_T1T2}
\alias{methylationGLMM_T1T2}
\title{Run methylationGLMM_T1T2.R as an external script from inst/script/.R}
\usage{
methylationGLMM_T1T2(
  inputPheno = "rData/preprocessingPheno/mergeData/phenoBetaT1T2.RData",
  outputLogs = "logs/",
  outputRData = "rData/methylationGLMM_T1T2/models",
  outputPlots = "figures/methylationGLMM_T1T2",
  personVar = "person",
  timeVar = "Timepoint",
  phenotypes =
    "DASS_Depression,DASS_Anxiety,DASS_Stress,PCL5_TotalScore,MHCSF_TotalScore,BRS_TotalScore",
  covariates = "Sex,Age,Ethnicity,TraumaDefinition,Leukocytes,Epithelial.cells",
  factorVars = "Sex,Ethnicity,TraumaDefinition,Timepoint",
  lmeLibs = "lme4,lmerTest",
  prsMap = NULL,
  libPath = NULL,
  cpgPrefix = "cg",
  cpgLimit = NA,
  nCores = 32,
  summaryPval = NA,
  plotWidth = 2000,
  plotHeight = 1000,
  plotDPI = 150,
  interactionTerm = NULL,
  saveSignificantInteractions = TRUE,
  significantInteractionDir = "preliminaryResults/cpgs/methylationGLMM_T1T2",
  significantInteractionPval = 0.05,
  saveTxtSummaries = TRUE,
  chunkSize = 10000,
  summaryTxtDir = "preliminaryResults/summary/methylationGLMM_T1T2/lmer",
  fdrThreshold = 0.05,
  padjmethod = "fdr",
  annotationPackage = "IlluminaHumanMethylationEPICv2anno.20a1.hg38",
  annotationCols =
    "Name,chr,pos,UCSC_RefGene_Group,UCSC_RefGene_Name,Relation_to_Island,GencodeV41_Group",
  annotatedLMEOut = "data/methylationGLMM_T1T2"
)
}
\description{
Run methylationGLMM_T1T2.R as an external script from inst/script/.R
}
